move src/lib/semconv.ts to src/semconv.ts
