+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/AwsEc2DetectorSync.ts
@@ -34,7 +34,7 @@ import {
   ATTR_HOST_NAME,
   CLOUD_PROVIDER_VALUE_AWS,
   CLOUD_PLATFORM_VALUE_AWS_EC2,
-} from '../lib/semconv';
+} from '../semconv';
 import * as http from 'http';
 
 /**
