+++ detectors/node/opentelemetry-resource-detector-aws/src/semconv.ts
